add the config parsing
