+++ Cargo.toml
@@ -8,5 +8,6 @@ clap = { version = "4.5.31", features = ["derive"] }
 futures = "0.3.31"
 libp2p = { version = "0.55.0", features = ["tokio", "yamux", "cbor", "serde", "noise", "identify", "tcp", "kad", "gossipsub", "request-response", "macros"] }
 serde = { version = "1.0.218", features = ["derive"] }
+serde_yaml = "0.9.34"
 tokio = { version = "1.43.0", features = ["full"] }
 tracing = "0.1.41"
