+++ src/consensus/mod.rs
@@ -1,2 +1,3 @@
-pub mod message;
-pub mod network;
+mod message;
+mod network;
+pub mod engine;
